requirements.txt
absl-py==0.9.0
alembic==1.3.3
asn1crypto==0.24.0
astor==0.8.1
backports.functools-lru-cache==1.6.1
backports.weakref==1.0.post1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.1
enum34==1.1.6
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
mahotas==1.4.9
Mako==1.1.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mock==3.0.5
networkx==2.2
numpy==1.16.6
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==2.3.2
Pillow==6.2.2
progressbar==2.5
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyWavelets==1.0.3
pyxdg==0.25
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.14.5
scikit-learn==0.20.4
scipy==1.2.2
SecretStorage==2.3.1
six==1.13.0
SQLAlchemy==1.3.13
subprocess32==3.5.4
tensorboard==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
